More work on setting up remotes
