+++ Cargo.toml
@@ -8,6 +8,7 @@ edition = "2018"
 derivative = "1.0.3"
 futures-preview = { version = "0.3.0-alpha.18", features = ["async-await", "nightly"] }
 log = "0.4.8"
+num-derive = "0.2.5"
 thespian-derive = { path = "./thespian-derive" }
 runtime = "0.3.0-alpha.7"
